uiux: Overhaul sidebar and use it for sitewide navigation. (#488)
The 'drawer' is now referred to as the 'sidebar.' The corporation links are added to the beginning of the sidebar. The tags are divided into two groups: general and specific. The sidebar is used on the platforms page now, too, which just places the alternatives right at the beginning of the main section of the page.